Some changes in onboarding to feature branch 1
