+++ ArbuzClone/View/HomeView/HomeView.swift
@@ -1,0 +1,18 @@
+//
+//  HomeVIew.swift
+//  ArbuzClone
+//
+//  Created by Rauan on 20.05.2024.
+//
+
+import SwiftUI
+
+struct HomeView: View {
+    var body: some View {
+        Text("home")
+    }
+}
+
+#Preview {
+    HomeView()
+}
